+++ pages/lessons.tsx
@@ -1,17 +1,17 @@
 import Head from "next/head";
-import styles from "../styles/Home.module.css";
+import styles from "../styles/Lessons.module.css";
 import Lessons from "../components/Lessons/Lessons";
 
 export default function LessonsPage() {
   return (
-    <div className={styles.container}>
+    <>
       <Head>
         <title>Guitar Chords – Lessons</title>
         <link rel="icon" href="/favicon.ico" />
       </Head>
-      <main>
+      <main className={styles.main}>
         <Lessons />
       </main>
-    </div>
+    </>
   );
 }
