+++ styles/Lessons.module.css
@@ -1,20 +1,12 @@
-.container {
+.main {
   background: rgb(255, 216, 0);
   background: linear-gradient(
     10deg,
     rgba(255, 216, 0, 1),
     rgba(255, 245, 188, 1) 100%
   );
+  padding: 0;
   height: 100vh;
-  padding: 0;
-}
-.main {
-  padding: 0;
-  height: 100%;
-  display: grid;
-  grid-template-columns: auto;
-  justify-items: center;
-  align-items: start;
 }
 .fullHeader {
   padding-top: 2rem;
